# backend/api/openapi/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Digital Store API",
    "description": "A digital product e-commerce system API providing endpoints for product browsing, shopping cart management, order processing, payment handling, and admin operations.",
    "version": "1.0",
    "contact": {
      "name": "API Support",
      "email": "support@digitalstore.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "schemes": ["http", "https"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Type \"Bearer\" followed by a space and JWT token."
    }
  },
...
  "paths": {
...
    "/orders": {
      "post": {
        "tags": ["orders"],
        "summary": "Create order",
        "description": "Create a new order from cart items. Supports both authenticated users and guests. Response time < 500ms.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/CreateOrderRequest" }
          }
        ],
        "responses": {
          "201": { "description": "Order created" },
          "400": { "description": "Invalid request or empty cart" }
        }
      },
      "get": {
        "tags": ["orders"],
        "summary": "List user orders",
        "description": "Get paginated list of orders for the authenticated user",
        "security": [{"BearerAuth": []}],
        "produces": ["application/json"],
        "responses": {
          "200": { "description": "Order list" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/orders/track": {
      "post": {
        "tags": ["orders"],
        "summary": "Track order",
        "description": "Track an order by order number and email (for guest orders)",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["order_number", "email"],
              "properties": {
                "order_number": { "type": "string", "example": "ORD-20240101-001" },
                "email": { "type": "string", "format": "email", "example": "user@example.com" }
              }
            }
          }
        ],
        "responses": {
          "200": { "description": "Order details" },
          "404": { "description": "Order not found" }
        }
      }
    },
...
  },
  "definitions": {
...
    "CreateOrderRequest": {
      "type": "object",
      "required": ["shipping_address", "payment_method"],
      "properties": {
        "shipping_address": { "type": "string", "example": "123 Main St, City, Country" },
        "payment_method": { "type": "string", "enum": ["online", "transfer"], "example": "online" },
        "guest_email": { "type": "string", "format": "email", "example": "guest@example.com" },
        "guest_name": { "type": "string", "example": "Guest User" },
        "guest_phone": { "type": "string", "example": "+1234567890" }
      }
    },
...
  }
}